perf(kubernetes): change KubernetesNamedAccountCredentials::getNamespaces to not make live calls to a kubernetes cluster (#5363)
* refactor(kubernetes): change KubernetesCredentials::getDeclaredNamespaces to return an ImmutableList

since the lists it returns are already (at least pre-filtering) immutable

* refactor(kubernetes): separate KubernetesCredentials::filterNamespaces into its own method

to prepare for an additional use

* perf(kubernetes): change KubernetesNamedAccountCredentials::getNamespaces to not make live calls to a kubernetes cluster

KubernetesNamedAccountCredentials::getNamespaces was already a best effort call since it could return an empty list in
case of an error communicating with the kubernetes cluster.  With this change, it doesn't attempt to communicate with
the kubernetes cluster and returns the last cached valued (or an empty list if there's nothing in the cache).  This can
dramatically speed up the /credentials endpoint if it's called soon after clouddriver starts, before a caching agent has
made a live call to the cluster to retrieve the list of namespaces.

* log a warning when the cache is empty

* no need for kube config file when service account is true

(cherry picked from commit 70f3fbe415a0d8b7d04436f1ef36503fbc359c5e)